Move a lot of files and a lot of features to the Bundle structure
