+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 Extension "Contao Form Conditional Notifications" for Contao Open Source CMS
 ========
 
+2.x - Soon
+- Compatibility update with Contao 5.x
+
+1.0.0 - 2025-01-24
+- Compatibility update with Contao 4.13
+
 0.1.0 - 2018-09-10
 - Init git repository
 - Add the main features of the extension
