+++ README.md
@@ -3,38 +3,34 @@ Extension "Contao Form Conditional Notifications" for Contao Open Source CMS
 
 Extends the way Notification Center deals with Contao Forms by allow user to configure a Notification, depending on Form Field Values.
 
+After a long time with no updates, this bundle is now compatible with Contao 4.13 & PHP 8. We will also work on the compatibility with Contao 5.3 and the Notification Center 2 later this year.
+
 Functionnalities
 -------------------
- * Configure Notifications as many as you want.
+ * Send notifications based on form data
+ * Adjust notifications tokens based on form data
 
 System requirements
 -------------------
- * Contao 3.*
- * Contao 4.* (Manager Edition)
-
+ * Contao 4.13 (Manager Edition)
 
 Installation
 ------------
 Clone the extension from Packagist (Contao Manager)
-or
-Download and paste the folder in system/modules of your Contao Website.
 
 Documentation
 -------------
  * [Change log][1]
  * [Git repository][2]
 
-
 License
 -------
 This extension is licensed under the terms of the Apache License 2.0. The full license text is
 available in the main folder.
 
-
 Getting support
 ---------------
 Visit the [support page][3] to submit an issue or just get in touch :)
-
 
 Installing from Git
 -------------------
