--- config/autoload.ini
@@ -1,19 +1,0 @@
-;;
-; List modules which are required to be loaded beforehand
-;;
-requires[] = "core"
-
-;;
-; Configure what you want the autoload creator to register
-;;
-register_namespaces = true
-register_classes    = true
-register_templates  = true
-
-;;
-; Override the default configuration for certain sub directories
-;;
-[vendor/*]
-register_namespaces = false
-register_classes    = false
-register_templates  = false
