--- config/autoload.php
@@ -1,15 +1,0 @@
-<?php
-
-/**
- * Form Conditional Notifications Extension for Contao Open Source CMS
- *
- * Copyright (c) 2018 Web ex Machina
- *
- * @author Web ex Machina <https://www.webexmachina.fr>
- */
-
-/**
- * Register PSR-0 namespace
- */
-if (class_exists('NamespaceClassLoader'))
-    NamespaceClassLoader::add('WEM', 'system/modules/wem-contao-form-conditional-notifications/library');
