+++ src/Resources/contao/dca/tl_wem_form_conditional_notification_field.php
@@ -1,0 +1,136 @@
+<?php
+
+use WEM\FormConditionalNotificationsBundle\DataContainer\Field;
+
+/**
+ * Table tl_wem_form_conditional_notification_field
+ */
+$GLOBALS['TL_DCA']['tl_wem_form_conditional_notification_field'] = array
+(
+
+	// Config
+	'config' => array
+	(
+		'dataContainer'               => 'Table',
+		'enableVersioning'            => true,
+		'ptable'                      => 'tl_wem_form_conditional_notification',
+		'onload_callback'			  => array
+		(
+			array(Field::class, 'adjustPalette')
+		),
+		'sql' => array
+		(
+			'keys' => array
+			(
+				'id' => 'primary',
+				'pid' => 'index',
+				'field' => 'index'
+			)
+		)
+	),
+
+	// List
+	'list' => array
+	(
+		'sorting' => array
+		(
+			'mode'                    => 4,
+			'fields'                  => array('sorting'),
+			'panelLayout'             => 'filter;search,limit',
+			'headerFields'            => array('nc_notification', 'language'),
+			'child_record_callback'   => array(Field::class, 'listItems'),
+			'disableGrouping'		  => true
+		),
+		'global_operations' => array
+		(
+			'all' => array
+			(
+				'href'                => 'act=select',
+				'class'               => 'header_edit_all',
+				'attributes'          => 'onclick="Backend.getScrollOffset()" accesskey="e"'
+			)
+		),
+		'operations' => array
+		(
+			'edit' => array
+			(
+				'href'                => 'act=edit',
+				'icon'                => 'edit.gif'
+			),
+			'delete' => array
+			(
+				'href'                => 'act=delete',
+				'icon'                => 'delete.gif',
+				'attributes'          => 'onclick="if(!confirm(\'' . $GLOBALS['TL_LANG']['MSC']['deleteConfirm'] . '\'))return false;Backend.getScrollOffset()"'
+			),
+			'show' => array
+			(
+				'href'                => 'act=show',
+				'icon'                => 'show.gif'
+			)
+		)
+	),
+
+	// Palettes
+	'palettes' => array
+	(
+		'default'                     => '{general_legend},field,value',
+	),
+
+	// Fields
+	'fields' => array
+	(
+		'id' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL auto_increment"
+		),
+		'pid' => array
+		(
+			'foreignKey'              => 'tl_wem_form_conditional_notification.id',
+			'sql'                     => "int(10) unsigned NOT NULL default '0'",
+			'relation'                => array('type'=>'belongsTo', 'load'=>'lazy')
+		),
+		'sorting' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL default '0'"
+		),
+		'tstamp' => array
+		(
+			'sql'                     => "int(10) unsigned NOT NULL default '0'"
+		),
+
+		'field' => array
+		(
+			'inputType'               => 'select',
+			'options_callback'        => array(Field::class, 'getFormFields'),
+			'foreignKey'              => 'tl_form_field.label',
+			'eval'                    => array('mandatory'=>true, 'submitOnChange'=>true, 'chosen'=>true, 'tl_class'=>'w50 wizard'),
+			'sql'                     => "int(10) unsigned NOT NULL default '0'"
+		),
+		'value' => array
+		(
+			'exclude'                 => true,
+			'inputType'               => 'text',
+			'eval'                    => array('maxlength'=>255, 'tl_class'=>'w50'),
+			'sql'                     => "text NULL"
+		),
+	)
+);
+
+/**
+ * Provide miscellaneous methods that are used by the data configuration array.
+ *
+ * @author Web ex Machina <https://www.webexmachina.fr>
+ */
+class tl_wem_form_conditional_notification_field extends Backend
+{
+	/**
+	 * Import the back end user object
+	 */
+	public function __construct()
+	{
+		parent::__construct();
+
+		$this->import('BackendUser', 'User');
+	}
+}
