+++ src/Resources/contao/languages/fr/default.xlf
