+++ src/WEMFormConditionalNotificationsBundle.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace WEM\FormConditionalNotificationsBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * Configures the Contao bundle.
+ *
+ * @author Web ex Machina <https://www.webexmachina.fr>
+ */
+class WEMFormConditionalNotificationsBundle extends Bundle
+{
+}
